requirements.txt
langchain
langchain-openai
langchain-community
feedparser
faiss-cpu
beautifulsoup4
python-dotenv
tiktoken
